// src/platform/packages/shared/kbn-unified-histogram/types.d.ts
export interface UnifiedHistogramServices {
    data: DataPublicPluginStart;
    uiActions: UiActionsStart;
    uiSettings: IUiSettingsClient;
    fieldFormats: FieldFormatsStart;
    lens: LensPublicStart;
    storage: Storage;
    expressions: ExpressionsStart;
    capabilities: Capabilities;
    dataViews: DataViewsPublicPluginStart;
    fieldsMetadata?: FieldsMetadataPublicStart;
    analytics?: AnalyticsServiceStart;
}
// ...
export declare enum UnifiedHistogramSuggestionType {
    unsupported = "unsupported",
    lensSuggestion = "lensSuggestion",
    histogramForESQL = "histogramForESQL",
    histogramForDataView = "histogramForDataView"
}
export interface UnifiedHistogramSuggestionContext {
    suggestion: Suggestion | undefined;
    type: UnifiedHistogramSuggestionType;
}
export interface LensRequestData {
    dataViewId?: string;
    timeField?: string;
    timeInterval?: string;
    breakdownField?: string;
}
export declare enum LensVisServiceStatus {
    'initial' = "initial",
    'completed' = "completed"
}
export interface LensVisServiceState {
    status: LensVisServiceStatus;
    currentSuggestionContext: UnifiedHistogramSuggestionContext;
    visContext: UnifiedHistogramVisContext | undefined;
}
/**
 * Unified Histogram type for recreating a stored Lens vis
 */
export interface UnifiedHistogramVisContext {
    attributes: TypedLensByValueInput['attributes'];
    requestData: LensRequestData;
    suggestionType: UnifiedHistogramSuggestionType;
}
export interface UnifiedHistogramFetchParamsExternal {
    /**
     * The request adapter to use for the inspector
     */
    requestAdapter: RequestAdapter | undefined;
    /**
     * The current search session ID
     */
    searchSessionId: string | undefined;
    /**
     * The abort controller to use for requests
     */
    abortController?: AbortController;
    /**
     * The current data view
     */
    dataView: DataView;
    /**
     * The current query
     */
    query?: Query | AggregateQuery;
    /**
     * The current filters
     */
    filters?: Filter[];
    /**
     * The current time range
     */
    timeRange?: TimeRange;
    /**
     * The relative time range, used when timeRange is an absolute range (e.g. for edit visualization button)
     */
    relativeTimeRange?: TimeRange;
    /**
     * The ES|QL variables to use for the chart
     */
    esqlVariables?: ESQLControlVariable[];
    /**
     * The controls state to use for the chart
     */
    controlsState?: ControlPanelsState<OptionsListESQLControlState>;
    /**
     * The current columns
     */
    columns?: DatatableColumn[];
    /**
     * Preloaded data table sometimes used for rendering the chart in ES|QL mode
     */
    table?: Datatable;
    /**
     * The current breakdown field
     */
    breakdownField?: string;
    /**
     * The current time interval of the chart
     */
    timeInterval?: string;
    /**
     * The external custom Lens vis
     */
    externalVisContext?: UnifiedHistogramVisContext;
    /**
     * Callback to modify the default Lens vis attributes used in the chart
     */
    getModifiedVisAttributes?: (attributes: TypedLensByValueInput['attributes']) => TypedLensByValueInput['attributes'];
}
export type UnifiedHistogramFetchParams = Omit<UnifiedHistogramFetchParamsExternal, 'breakdownField' | 'timeInterval'> & {
    query: Query | AggregateQuery;
    filters: Filter[];
    timeRange: TimeRange;
    relativeTimeRange: TimeRange;
    esqlVariables: ESQLControlVariable[];
    lastReloadRequestTime: number;
    isESQLQuery: boolean;
    isTimeBased: boolean;
    columnsMap: Record<string, DatatableColumn> | undefined;
    breakdown: {
        field: DataViewField | undefined;
    } | undefined;
    timeInterval: string;
};